+++ spec/arachni/check/auditor_spec.rb
@@ -48,7 +48,7 @@ end
 describe Arachni::Check::Auditor do
     before :each do
         @opts = Arachni::Options.instance
-        @opts.audit.elements :links, :forms, :cookies, :headers, :jsons, :xmls
+        @opts.audit.elements Arachni::Page::ELEMENTS - [:link_templates]
 
         @opts.url = web_server_url_for( :auditor )
         @url      = @opts.url
@@ -68,6 +68,15 @@ describe Arachni::Check::Auditor do
         $audit_taint_called        = nil
         $audit_called              = nil
     end
+
+    element_classes = [
+        Arachni::Element::Link, Arachni::Element::Link::DOM,
+        Arachni::Element::Form, Arachni::Element::Form::DOM,
+        Arachni::Element::Cookie, Arachni::Element::Cookie::DOM,
+        Arachni::Element::Header, Arachni::Element::LinkTemplate,
+        Arachni::Element::LinkTemplate::DOM, Arachni::Element::JSON,
+        Arachni::Element::XML
+    ]
 
     let(:auditor) { AuditorTest.new( @framework ) }
     let(:url) { @url }
@@ -132,6 +141,25 @@ describe Arachni::Check::Auditor do
     end
 
     describe '.check?' do
+        context 'when elements have been provided' do
+            it 'restricts the check' do
+                page = Arachni::Page.from_data( url: url, body: 'stuff' )
+                page.stub(:has_script?) { true }
+                auditor.class.info[:elements] =
+                    element_classes + [Arachni::Element::Body, Arachni::Element::GenericDOM]
+
+                auditor.class.check?( page, Arachni::Element::GenericDOM ).should be_true
+                auditor.class.check?( page, Arachni::Element::Body ).should be_true
+
+                element_classes.each do |element|
+                    auditor.class.check?( page, element ).should be_false
+                end
+
+                auditor.class.check?( page, element_classes ).should be_false
+                auditor.class.check?( page, element_classes + [Arachni::Element::Body] ).should be_true
+            end
+        end
+
         context Arachni::Element::Body do
             before(:each) { auditor.class.info[:elements] = Arachni::Element::Body }
 
@@ -170,13 +198,6 @@ describe Arachni::Check::Auditor do
                 end
             end
         end
-
-        element_classes = [Arachni::Element::Link, Arachni::Element::Link::DOM,
-                           Arachni::Element::Form, Arachni::Element::Form::DOM,
-                           Arachni::Element::Cookie, Arachni::Element::Cookie::DOM,
-                           Arachni::Element::Header, Arachni::Element::LinkTemplate,
-                           Arachni::Element::LinkTemplate::DOM, Arachni::Element::JSON,
-                           Arachni::Element::XML ]
 
         element_classes.each do |element|
             context "when #{Arachni::OptionGroups::Audit}##{element.type.to_s.gsub( '_dom', '')}? is" do