+++ spec/support/factories/page.rb
@@ -1,7 +1,7 @@
 Factory.define :page do
     Arachni::Page.new(
         response: Factory[:response],
-        dom:      Factory[:dom_data],
-        metadata: { stuf: 'blah' }
-    )
+        dom:      Factory[:dom_data]
+    # Load all elements to populate metadata and the like but clear the cache.
+    ).tap(&:elements).tap(&:clear_cache)
 end